+++ ReferenceTests/src/tests/examples2d.jl
@@ -929,6 +929,397 @@ end
     fig
 end
 
+@reference_test "comparison of tricontourf and tricontour" begin
+    x = RNG.randn(100)
+    y = RNG.randn(100)
+    z = -sqrt.(x .^ 2 .+ y .^ 2) .+ 0.1 .* RNG.randn(100);
+    levels = 5;
+
+    f, ax, trf = tricontourf(x, y, z, levels = levels);
+    ax.title = "tricontourf and tricontour"
+    scatter!(x, y, color = z, strokewidth = 1, strokecolor = :black)
+    Colorbar(f[1, 2], trf);
+
+    # Add contour lines
+    trl = tricontour!(
+        ax, x, y, z,
+        levels = levels - 1,
+        colormap=:reds,
+        linewidth=3,  linestyle=:dashdot,
+        );
+    Colorbar(f[1, 3], trl.plots[2]);
+
+    f
+end
+
+@reference_test "comparison of tricontour and contour" begin
+
+    f = Figure()
+    ax1 = Axis(f[1, 1], title="contour and tricontour", aspect = DataAspect())
+
+    xs = LinRange(0, 10, 100)
+    ys = LinRange(0, 15, 100)
+    xs1 = [x for x in xs, y in ys]
+    ys1 = [y for x in xs, y in ys]
+    zs1 = [cos(x) * sin(y) for x in xs, y in ys]
+    
+    levels = -1:0.1:1
+    contour!(ax1, xs1, ys1, zs1, levels=levels)
+    
+    mask_region(x, y) = ((x-5)/4)^2 + ((y-7)/6)^2 < 1
+    
+    xs2 = [x for x in xs, y in ys if mask_region(x,y)]
+    ys2 = [y for x in xs, y in ys if mask_region(x,y)]
+    zs2 = [cos(x) * sin(y) for x in xs, y in ys if mask_region(x,y)]
+
+    tr = tricontour!(
+        ax1, xs2[:], ys2[:], zs2[:], levels=levels,
+        color=:black, linestyle = :dash
+        )
+    f
+end
+
+@reference_test "tricontour with :relative and :normal level modes" begin
+    
+    x = RNG.randn(100)
+    y = RNG.randn(100)
+    z = x .* y
+    
+    f = Figure()
+    ax1 = Axis(f[1,1], title="Normal `levels` (values of z)")
+    ax2 = Axis(f[1,2], title="Relative `levels` (percentiles of z)")
+    levels = [0.1, 0.25, 0.5, 0.75, 1.0]
+
+    tr_abs_levels = tricontour!(
+        ax1, x, y, z, colormap = :viridis, linewidth = 3, alpha=0.5,
+        levels = levels, mode=:normal
+        )
+    tr_rel_levels = tricontour!(
+        ax2, x, y, z, colormap = :viridis, linewidth = 3, alpha=0.5,
+        levels = levels, mode=:relative
+        )
+
+    scatter!(
+        ax1, x, y, color = z, colormap = :viridis,
+        strokecolor = :black, strokewidth = 2
+        )
+    scatter!(
+        ax2, x, y, color = z, colormap = :viridis,
+        strokecolor = :black, strokewidth = 2
+        )
+    
+    Colorbar(f[2, 1], tr_abs_levels.plots[2], vertical = false)
+    Colorbar(f[2, 2], tr_rel_levels.plots[2], vertical = false)
+
+    f
+end
+
+@reference_test "tricontour level options (no labels)" begin
+
+    xs = LinRange(-5, 5, 100);
+    ys = LinRange(-5, 5, 100);
+    X = [x for x in xs, y in ys][:];
+    Y = [y for x in xs, y in ys][:];
+
+    # Implicit shape of a heart: f(x, y) = k
+    f = (x,y) -> x^2 + (y - (x^2)^(1/3))^2
+    Z = f.(X, Y)
+
+    fig = Figure()
+
+    # Add a supertitle
+    supertitle = Label(
+        fig[0, :], "Tricontour level options (no labels)",
+        fontsize=18, font=:bold
+        )
+    
+    ax1 = Axis(fig[1,1], title="Default (10 levels) [no labels]")
+    ax2 = Axis(fig[1,3], title="Integer levels (5 levels)")
+    ax3 = Axis(fig[2,1], title="Vector of 4 floats as levels")
+    ax4 = Axis(fig[2,3], title="Vector with single value as level")
+
+    tr1 = tricontour!(ax1, X, Y, Z)
+    Colorbar(fig[1, 2], tr1.plots[2])
+
+    tr2 = tricontour!(
+        ax2, X, Y, Z, levels = 5
+        )
+    Colorbar(fig[1, 4], tr2.plots[2])
+
+    tr3 = tricontour!(
+        ax3, X, Y, Z, levels = [0.5, 1.0, 2.0, 3.0]
+        )
+    Colorbar(fig[2, 2], tr3.plots[2])
+
+    tr4 = tricontour!( ax4, X, Y, Z, levels = [2.0] )
+    Colorbar(fig[2, 4], tr4.plots[2])
+    
+    fig
+    
+end
+
+@reference_test "tricontour level options (with labels)" begin
+
+    xs = LinRange(-5, 5, 100);
+    ys = LinRange(-5, 5, 100);
+    X = [x for x in xs, y in ys][:];
+    Y = [y for x in xs, y in ys][:];
+
+    # Implicit shape of a heart: f(x, y) = k
+    f = (x,y) -> x^2 + (y - (x^2)^(1/3))^2
+    Z = f.(X, Y)
+    
+    fig = Figure()
+    ax1 = Axis(fig[1,1], title="Default (10 levels) [with labels]")
+    ax2 = Axis(fig[1,3], title="Integer levels (5 levels)")
+    ax3 = Axis(fig[2,1], title="Vector of 4 floats as levels")
+    ax4 = Axis(fig[2,3], title="Vector with single value as level")
+
+    tr1 = tricontour!(ax1, X, Y, Z, labels = true)
+    Colorbar(fig[1, 2], tr1.plots[2])
+
+    tr2 = tricontour!(
+        ax2, X, Y, Z,
+        levels = 5, labels = true
+        )
+    Colorbar(fig[1, 4], tr2.plots[2])
+
+    tr3 = tricontour!(
+        ax3, X, Y, Z,
+        levels = [0.5, 1.0, 2.0, 3.0], labels = true
+        )
+    Colorbar(fig[2, 2], tr3.plots[2])
+
+    tr4 = tricontour!(
+        ax4, X, Y, Z,
+        levels = [2.0], labels = true
+        )
+    Colorbar(fig[2, 4], tr4.plots[2])
+    
+    fig
+    
+end
+
+@reference_test "tricontour linestyle format options" begin
+
+    xs = LinRange(-3, 3, 100);
+    ys = LinRange(-3, 3, 100);
+    X = [x for x in xs, y in ys][:];
+    Y = [y for x in xs, y in ys][:];
+
+    levels = collect(-5:10)
+
+    # Contours f(x, y) = k
+    f = (x,y) -> 0.25 * ( 5*x^2 + y^2 -4 ) * ( x^2 + 5*y^2 -4 )
+    Z = f.(X, Y)
+    
+    fig = Figure()
+    ax1 = Axis(fig[1,1], title="Linewidth 2.0 + linecap :round")
+    ax2 = Axis(fig[1,3], title="Linewidth 3.0 + alpha 0.5")
+    ax3 = Axis(fig[2,1], title="Linestyle (:dot, :loose)")
+    ax4 = Axis(fig[2,3], title="Linewidth 2.0 + joinstyle :bevel")
+
+
+    tr1 = tricontour!(
+        ax1, X, Y, Z, levels = levels,
+        linewidth = 2.0, linecap = :round
+        )
+    Colorbar(fig[1, 2], tr1.plots[2])
+
+    tr2 = tricontour!(
+        ax2, X, Y, Z, levels = levels,
+        linewidth = 3, alpha = 0.5
+        )
+    Colorbar(fig[1, 4], tr2.plots[2])
+
+    tr3 = tricontour!(
+        ax3, X, Y, Z, levels = levels,
+        linestyle = (:dot, :loose)
+        )
+    Colorbar(fig[2, 2], tr3.plots[2])
+
+    tr4 = tricontour!(
+        ax4, X, Y, Z, levels = levels,
+        linewidth = 2.0, joinstyle=:bevel
+        )
+
+    fig
+    
+end
+
+@reference_test "tricontour color options" begin
+
+    xs = LinRange(-4, 4, 100);
+    ys = LinRange(-4, 4, 100);
+    X = [x for x in xs, y in ys][:];
+    Y = [y for x in xs, y in ys][:];
+
+    levels = [-0.5, 0, 0.5]
+
+    # Contours f(x, y) = k
+    f = (x,y) -> cos( 20 * ( atan((x-1)/y ) + atan(y, x+1 ) ) )
+    Z = f.(X, Y)
+    
+    fig = Figure()
+    ax1 = Axis(fig[1,1], title="Default")
+    ax2 = Axis(fig[1,3], title="Continuous colormap")
+    ax3 = Axis(fig[2,1], title="Colormap :reds")
+    ax4 = Axis(fig[2,3], title="Color :black")
+
+
+    tr1 = tricontour!(
+        ax1, X, Y, Z, levels = levels,
+        )
+    Colorbar(fig[1, 2], tr1.plots[2])
+
+    tr2 = tricontour!(
+        ax2, X, Y, Z, levels = levels,
+        discretize_colormap = false
+        )
+    Colorbar(fig[1, 4], tr2.plots[2])
+
+    tr3 = tricontour!(
+        ax3, X, Y, Z, levels = levels,
+        colormap = :reds
+        )
+    Colorbar(fig[2, 2], tr3.plots[2])
+
+    tr4 = tricontour!(
+        ax4, X, Y, Z, levels = levels,
+        color = :black
+        )
+
+    fig
+    
+end
+
+@reference_test "tricontour label format options" begin
+
+    xs = LinRange(-4, 4, 100);
+    ys = LinRange(-4, 4, 100);
+    X = [x for x in xs, y in ys][:];
+    Y = [y for x in xs, y in ys][:];
+
+    b = [0.8, 1, 1.1, 1.2, 1.4, 1.6]
+    levels = b.^4
+
+    round.(levels.^(0.25); digits=1)
+
+    # Cassini ovals: f_cassini(x, y, a) = b^4
+    f_cassini = (x, y, a) -> ( x^2 + y^2 )^2 - 2*a^2*( x^2 - y^2) + a^4
+    f = (x, y) -> f_cassini(x, y, 1)
+    Z = f.(X, Y)
+
+    # Inline function for label strings
+    custom_label_formatter = (level::Real) -> begin
+        b = round( level^0.25; digits = 2 )
+        "b= " * string(isinteger(b) ? round(Int, b) : b)
+    end
+    
+    fig = Figure()
+    ax1 = Axis(fig[1,1], title="Default labels")
+    ax2 = Axis(fig[1,3], title="Custom labelformatter, inline")
+    ax3 = Axis(fig[2,1], title="Color blue, labelcolor black α=0.5")
+    ax4 = Axis(fig[2,3], title="Label font and size")
+
+
+    tr1 = tricontour!(
+        ax1, X, Y, Z, levels = levels, labels = true
+        )
+    Colorbar(fig[1, 2], tr1.plots[2])
+
+    tr2 = tricontour!(
+        ax2, X, Y, Z, levels = levels, labels = true,
+        labelformatter = custom_label_formatter,
+        inline=false, alpha=0.3
+        )
+    Colorbar(fig[1, 4], tr2.plots[2])
+
+    tr3 = tricontour!(
+        ax3, X, Y, Z, levels = levels, labels = true,
+        color = :blue, labelcolor = (:black, 0.5)
+        )
+
+    tr4 = tricontour!(
+        ax4, X, Y, Z, levels = levels, labels = true,
+        labelfont = :bold_italic, labelsize = :8
+        )
+    Colorbar(fig[2, 4], tr4.plots[2])
+    fig
+    
+end
+
+@reference_test "tricontour and triplot with Delaunay Triangulation" begin
+    
+    using DelaunayTriangulation
+    
+    outer = [
+        (0.0,0.0),(2.0,1.0),(4.0,0.0),
+        (6.0,2.0),(2.0,3.0),(3.0,4.0),
+        (6.0,6.0),(0.0,6.0),(0.0,0.0)
+    ];
+    inner = [
+        (1.0,5.0),(2.0,4.0),(1.01,1.01),
+        (1.0,1.0),(0.99,1.01),(1.0,5.0)
+    ];
+    boundary_points = [[outer], [inner]]
+    boundary_nodes, points = convert_boundary_points_to_indices(boundary_points)
+    tri = triangulate(points; boundary_nodes = boundary_nodes)
+    
+    refine!(tri; max_area=1e-3*get_area(tri))
+    
+    f, ax, tr = triplot(
+        tri, show_constrained_edges = true, constrained_edge_linewidth = 3,
+        constrained_edge_color = :black,
+        show_convex_hull = true, strokecolor=(:gray, 0.2), strokewidth = 2
+        )
+    
+    x = [p[1] for p in tri.points];
+    y = [p[2] for p in tri.points];
+    z = @. (sin(x)-3)^2 + (cos(y)-3)^2 - 10;
+    
+    tr = tricontour!(
+        ax, tri, z, labels=true, levels = collect(-2:2:16),
+        linewidth=3.0, labelsize=12, colormap = :cividis, labelcolor=:red,)
+    
+    Colorbar(f[1,2] ,tr.plots[2])
+    ax.title[]= L"f(x,y) = (sin(x) - 3)^2 + (cos(y) - 3)^2 - 10"
+
+    f
+end
+
+
+@reference_test "tricontour observables are updated" begin
+
+# Set up
+npoints = 500  # Number of random points
+x = 4 .* RNG.rand(npoints) .- 2 # Random x in [-2, 2]
+y = 4 .* RNG.rand(npoints) .- 2 # Random y in [-2, 2]
+
+# Function to compute z values with parameter `a`
+a = Observable(0.0)
+z = @lift ($a .* y - x .^ 2) .* ($a .* x - y .^ 2)
+
+# Create a reactive title using `lift`
+b = @lift round($a; digits=2)
+title_text = @lift "( $($b) * y - x^2 ) * ( $($b)*x - y^2 )"
+
+
+# Set up figure
+fig = Figure()
+ax = Axis(fig[1, 1], xlabel = "x", ylabel = "y",title = title_text)
+scatter!(ax, x, y, alpha=0.3)
+levels = collect(-5:1:10)
+contour_plot = tricontour!(ax, x, y, z, levels= levels, labels=true)
+Colorbar(fig[1,2], contour_plot.plots[2])
+
+# Update observable parameter 
+a[] = 2.0
+
+fig
+end
+
+
 @reference_test "trimspine" begin
     with_theme(Axis = (limits = (0.5, 5.5, 0.3, 3.4), spinewidth = 8, topspinevisible = false, rightspinevisible = false)) do
         f = Figure(size = (800, 800))