+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added `font` attribute and fixed faulty selection in `scatter`. Scatter fonts can now be themed with `markerfont`. [#4832](https://github.com/MakieOrg/Makie.jl/pull/4832)
 - Fixed categorical `cgrad` interpolating at small enough steps [#4858](https://github.com/MakieOrg/Makie.jl/pull/4858)
+- Added `tricontour` plot, which plots contour lines, similar to `tricontourf` [#4795](https://github.com/MakieOrg/Makie.jl/pull/4795).
 
 ## [0.22.2] - 2025-02-26